Staging

+++ Core/Controls/NotificationDirector.cs
@@ -5,8 +5,10 @@ using System.Text;
 using System.Threading.Tasks;
 using Core.Boundaries;
 using Core.Entities;
+using Core.Notifications;
 
 using static Core.Entities.Psijic;
+using static Core.Entities.Artificer;
 
 namespace Core.Controls
 {
@@ -20,7 +22,49 @@ namespace Core.Controls
 
 		#region Operations
 
-		public async Task<List<TelegramShard>> GetTelegramsAsync(long userId)
+		public async Task<NotificationPreferencesShard> GetNotificationPreferencesAsync(long userId)
+		{
+			var user = await GetUserAsync(userId);
+			var profile = await user.NotificationProfile;
+
+			return new(profile.NotificationId, profile.SocialInvitations, profile.CompanionActivity,
+				profile.GatheringReminders, profile.GatheringActivity, profile.GatheringDiscovery);
+		}
+
+        public async Task UpdateNotificationPreferencesAsync(long userId,
+            bool? socialInvitations = null, bool? companionActivity = null,
+			bool? gatheringReminders = null, bool? gatheringActivity = null,
+			bool? gatheringDiscovery = null)
+		{
+			var user = await GetUserAsync(userId);
+
+            List<(string Property, object Value)> edits = new();
+
+            if (IsNotNull(socialInvitations))
+			{
+                edits.Add((nameof(NotificationProfile.SocialInvitations), socialInvitations.Value));
+            }
+            if (IsNotNull(companionActivity))
+			{
+                edits.Add((nameof(NotificationProfile.CompanionActivity), companionActivity.Value));
+            }
+            if (IsNotNull(gatheringReminders))
+			{
+                edits.Add((nameof(NotificationProfile.GatheringReminders), gatheringReminders.Value));
+            }
+            if (IsNotNull(gatheringActivity))
+			{
+                edits.Add((nameof(NotificationProfile.GatheringActivity), gatheringActivity.Value));
+            }
+            if (IsNotNull(gatheringDiscovery))
+			{
+                edits.Add((nameof(NotificationProfile.GatheringDiscovery), gatheringDiscovery.Value));
+            }
+
+			await Telegrams.UpdateNotificationProfileAsync(user.Id, edits);
+		}
+
+        public async Task<List<TelegramShard>> GetTelegramsAsync(long userId)
 		{
 			var user = await GetUserAsync(userId);
 
@@ -32,6 +76,7 @@ namespace Core.Controls
 			var user = await GetUserAsync(userId);
 
 			var telegrams = await Telegrams.GetTelegramsAsync(user.Id);
+
 			foreach (var telegram in telegrams)
 			{
 				try
@@ -46,19 +91,32 @@ namespace Core.Controls
 		{
 			var user = await GetUserAsync(userId);
 
+			var telegrams = await Telegrams.GetTelegramsAsync(user.Id);
+
 			foreach (var id in telegramIds)
 			{
-				try
+				// Check if user owns telegram
+				// TODO Potential flag for abuse
+				if (telegrams.Exists(t => t.Id.Equals(id)))
 				{
-					await Telegrams.DeleteTelegramAsync(id);
+					try
+					{
+						await Telegrams.DeleteTelegramAsync(id);
+					}
+					catch { }
 				}
-				catch { }
+
 			}
 		}
 
 		#endregion
 
 		#region Favours
+
+		internal async Task<NotificationProfile> RequestNotificationProfileAsync(User user)
+		{
+			return await Telegrams.GetNotificationProfileAsync(user.Id);
+		}
 
 		internal async Task PostTelegramAsync(User user, User notifier, TelegramMessage message, string context)
 		{
@@ -69,9 +127,46 @@ namespace Core.Controls
 			await Telegrams.SaveTelegramAsync(user.Id, notifier.Id, Time, message, context);
 		}
 
-		internal async Task NotifyUserAsync(User user, NotificationGroup group, string title, string message, string collapseId = "")
+		internal async Task<string> NotifyUserAsync(User user, CanaryNotification notification, DateTimeOffset? notifyAt = null)
 		{
-            await Terminal.NotificationService.PushNotification(user.NotificationId.ToString(), group, title, message, collapseId);
+			string notificationId;
+
+			if (IsNotNull(notifyAt))
+			{
+                notificationId = await Terminal.NotificationService.ScheduleNotification(notification, notifyAt.Value, await user.NotificationProfile);
+			}
+			else
+			{
+				notificationId = await Terminal.NotificationService.DispatchNotification(notification, await user.NotificationProfile);
+            }
+
+			return notificationId;
+        }
+
+		internal async Task<string> NotifyUsersAsync(CanaryNotification notification, DateTimeOffset? notifyAt = null, params User[] users)
+		{
+			var profiles = await Task.WhenAll(users.Select(async user => await user.NotificationProfile));
+
+			string notificationId;
+
+			if (IsNotNull(notifyAt))
+			{
+				notificationId = await Terminal.NotificationService.ScheduleNotification(notification, notifyAt.Value, profiles);
+			}
+			else
+			{
+				notificationId = await Terminal.NotificationService.DispatchNotification(notification, profiles);
+			}
+
+			return notificationId;
+		}
+
+		internal async Task CancelScheduledNotifications(params string[] notificationIds)
+		{
+			foreach (var id in notificationIds)
+			{
+				await Terminal.NotificationService.CancelNotification(id);
+			}
 		}
 
 		#endregion
